Improving layout for blockquote - closes #4
In older posts such as /blog/2006-02-26-framework/ this is still not looking perfect.

The reason is that Bootstrap seems to have a different idea bout what <blockquote> is used to. Rather than using it to quote lengthy sections from other places (blog posts, literature, documentation, …) Bootstrap intends to use it to lift out and quote short sections of the text itself, e.g. in an interview. I don’t want to change this further, as we are better of adhering to Bootstrap way of thinking in the future, rather than patching Bootstrap to work with old blog posts.